Move all monkey patches to their own file
