+++ bot/__init__.py
@@ -5,8 +5,7 @@ from typing import TYPE_CHECKING
 
 from discord.ext import commands
 
-from bot import log, typing
-from bot.command import Command
+from bot import log, monkey_patches
 
 if TYPE_CHECKING:
     from bot.bot import Bot
@@ -17,11 +16,11 @@ log.setup()
 if os.name == "nt":
     asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())
 
-typing.patch_typing()
+monkey_patches.patch_typing()
 
 # Monkey-patch discord.py decorators to use the Command subclass which supports root aliases.
 # Must be patched before any cogs are added.
-commands.command = partial(commands.command, cls=Command)
-commands.GroupMixin.command = partialmethod(commands.GroupMixin.command, cls=Command)
+commands.command = partial(commands.command, cls=monkey_patches.Command)
+commands.GroupMixin.command = partialmethod(commands.GroupMixin.command, cls=monkey_patches.Command)
 
 instance: "Bot" = None  # Global Bot instance.
